Created ProgressBar component, refact. HuntXpBar
Refactored commonalities of progress bars into
ProgressBar component, began creating components
for the separate stat progress bars
